Pint

+++ app/DynamicLoading.php
@@ -4,22 +4,19 @@ namespace App;
 
 trait DynamicLoading
 {
-	protected $total_photos = 0;
+    protected $total_photos = 0;
 
-	protected $months = ['', 'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
+    protected $months = ['', 'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
 
-	protected $latlong = [];
+    protected $latlong = [];
 
-	protected $photoCount = 0;
+    protected $photoCount = 0;
 
-	/**
-	 *
-	 */
-	protected function getInitialPhotoLatLon ($photoData)
-	{
-		$lat = (double)$photoData->lat;
-		$lon = (double)$photoData->lon;
-		$this->latlong[0] = $lat;
-		$this->latlong[1] = $lon;
-	}
+    protected function getInitialPhotoLatLon($photoData)
+    {
+        $lat = (float) $photoData->lat;
+        $lon = (float) $photoData->lon;
+        $this->latlong[0] = $lat;
+        $this->latlong[1] = $lon;
+    }
 }
